Indent control blocks in helm partial templates for readability
Also, add a template partial, `indent`, that indents all lines in a
string with an exact number of spaces. This allows us to include control
blocks with varying levels of indentation in partial templates like
"yugabyte.labels". When rendering the partial template, we can realign
all lines to the same indentation level with the new `indent` partial
template.